requirements.txt
pytest==7.1.3
requests~=2.28.1
allure-pytest==2.10.0
selene>=2.0.0b5
python-dotenv>=0.21.0
voluptuous~=0.13.1
pytest_voluptuous>=1.2.0
curlify>=2.2.1
pip==22.3

Appium-Python-Client~=2.6.1
selenium~=4.4.3
pydantic~=1.10.2